Adding more to benchmarking
